gui: use shorten_mixes_timestamps method
from Mix_view_common for mixes list timestamp formatting. Might be an
issue if we implement sorting by Played col as sorting should happen
with the original date and not the reformatted string. Same applies to
editing. FIXME the original date should be somewhere hidden in the data
to have it accessible when editing or sorting.