Fix ValueError with default discogs_sold_folder_id
as reported in issue #40.

The config.yaml setting discogs_sold_folder_id defaults to '' (empty
string), the comparision in create_collection_folder() tries to cast an
int from it which can only fail with a ValueError.

Now it's sanity-checked if we have a number we can compare with or fall
back to simply designate the sold state as 0 (not sold). Numbers as
strings in config.yaml will also work (typically users will put their
id between the empty string quotes instead of replacing the whole yaml
value with a real number).